Lesson 17 (Exercise)
Develop an application using Spring Boot that allows managing a list of students through a CRUD.
The application must provide RESTful endpoints for:
- Displaying all students.
- Querying a student by email.
- Creating a new student.
- Modifying a student's information (either completely or partially).
- Deleting a student by ID.

The resources of this application must be contained in a Java list, and initially, it must contain a maximum of 4 students with the attributes id, name, email, age, and course.
The routes of all endpoints must be unified.